# docs/api/backend.openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "ShopSteward Backend API — 设计契约（尚未实现）",
    "version": "0.2.0-draft",
    "description": "B0独立业务后端；B2为未来接口。Swagger展示不代表服务已实现。业务语义见../backend-development.md。"
  },
  "servers": [
    {
      "url": "http://localhost:8000",
      "description": "规划的本地backend地址，当前未启动"
    }
  ],
...
  "paths": {
...
    "/api/v1/alerts": {
      "get": {
        "operationId": "list_alerts",
        "tags": [
          "Alerts"
        ],
        "summary": "查询警报；默认仅返回未解除警报",
        "description": "查询警报；默认仅返回未解除警报",
        "x-phase": "B0",
        "x-implementation-status": "planned",
        "x-permissions": [
          "viewer"
        ],
        "security": [
          {
            "UserBearer": []
          }
        ],
        "parameters": [
          {
            "name": "store_id",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "description": "不透明资源 ID",
              "minLength": 1,
              "maxLength": 128
            }
          },
          {
            "name": "mission_id",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "description": "不透明资源 ID",
              "minLength": 1,
              "maxLength": 128
            }
          },
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "enum": [
                "OPEN",
                "ACKNOWLEDGED",
                "RESOLVED"
              ]
            }
          },
          {
            "name": "cursor",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 100,
              "default": 20
            }
          }
        ],
        "responses": {
          "200": {
            "description": "成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AlertList"
                },
                "example": {
                  "items": [
                    {
                      "id": "alert_001",
                      "store_id": "store_001",
                      "mission_id": "mission_001",
                      "sku_id": "sku_001",
                      "type": "STOCKOUT_RISK",
                      "severity": "WARNING",
                      "status": "OPEN",
                      "summary": "预计缺货40件",
                      "facts": {
                        "shortage_qty": 40
                      },
                      "state_version": 7,
                      "related_plan_id": "plan_001",
                      "first_seen_at": "2026-09-06T00:00:00Z",
                      "last_seen_at": "2026-09-06T00:00:00Z",
                      "resolved_at": null
                    }
                  ],
                  "next_cursor": null
                }
              }
            }
          },
          "401": {
            "description": "无有效身份",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "403": {
            "description": "无对应权限",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "资源不存在或不可见",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "422": {
            "description": "字段或业务约束无效",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "429": {
            "description": "超过频率限制",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "500": {
            "description": "内部错误",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "503": {
            "description": "服务暂不可用",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/alerts/{alert_id}/acknowledgement": {
      "post": {
        "operationId": "acknowledge_alert",
        "tags": [
          "Alerts"
        ],
        "summary": "确认知晓，不能消除风险或批准采购",
        "description": "确认知晓，不能消除风险或批准采购",
        "x-phase": "B0",
        "x-implementation-status": "planned",
        "x-permissions": [
          "operator"
        ],
        "security": [
          {
            "UserBearer": []
          }
        ],
        "parameters": [
          {
            "name": "alert_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "description": "不透明资源 ID",
              "minLength": 1,
              "maxLength": 128
            }
          },
          {
            "name": "Idempotency-Key",
            "in": "header",
            "required": true,
            "description": "同一逻辑命令网络重试复用；同键不同内容返回409。",
            "schema": {
              "type": "string",
              "minLength": 1,
              "maxLength": 128
            }
          }
        ],
        "responses": {
          "200": {
            "description": "成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Alert"
                },
                "example": {
                  "id": "alert_001",
                  "store_id": "store_001",
                  "mission_id": "mission_001",
                  "sku_id": "sku_001",
                  "type": "STOCKOUT_RISK",
                  "severity": "WARNING",
                  "status": "OPEN",
                  "summary": "预计缺货40件",
                  "facts": {
                    "shortage_qty": 40
                  },
                  "state_version": 7,
                  "related_plan_id": "plan_001",
                  "first_seen_at": "2026-09-06T00:00:00Z",
                  "last_seen_at": "2026-09-06T00:00:00Z",
                  "resolved_at": null
                }
              }
            }
          },
          "401": {
            "description": "无有效身份",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "403": {
            "description": "无对应权限",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "资源不存在或不可见",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "409": {
            "description": "版本、幂等或业务并发冲突",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "422": {
            "description": "字段或业务约束无效",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "429": {
            "description": "超过频率限制",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "500": {
            "description": "内部错误",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "503": {
            "description": "服务暂不可用",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/job-runs/{job_run_id}": {
      "get": {
        "operationId": "get_job_run",
        "tags": [
          "Scheduling"
        ],
        "summary": "查询单次异步任务结果",
        "description": "查询单次异步任务结果",
        "x-phase": "B0",
        "x-implementation-status": "planned",
        "x-permissions": [
          "viewer"
        ],
        "security": [
          {
            "UserBearer": []
          }
        ],
        "parameters": [
          {
            "name": "job_run_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "description": "不透明资源 ID",
              "minLength": 1,
              "maxLength": 128
            }
          }
        ],
        "responses": {
          "200": {
            "description": "成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobRun"
                },
                "example": {
                  "id": "job_001",
                  "mission_id": "mission_001",
                  "job_type": "check_mission",
                  "trigger_source": "MANUAL",
                  "status": "SUCCEEDED",
                  "attempt_count": 1,
                  "scheduled_for": "2026-09-06T00:00:00Z",
                  "available_at": "2026-09-06T00:00:00Z",
                  "started_at": "2026-09-06T00:00:00Z",
                  "finished_at": "2026-09-06T00:00:01Z",
                  "result": {
                    "check_status": "ANOMALY",
                    "summary": "完成检查，发现缺货并生成方案",
                    "references": [
                      {
                        "type": "plan",
                        "id": "plan_001",
                        "version": "1"
                      }
                    ],
                    "input_state_version": 7,
                    "output_state_version": 7
                  },
                  "error_code": null,
                  "last_error": null,
                  "request_id": "req_001"
                }
              }
            }
          },
          "401": {
            "description": "无有效身份",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "403": {
            "description": "无对应权限",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "资源不存在或不可见",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "422": {
            "description": "字段或业务约束无效",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "429": {
            "description": "超过频率限制",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "500": {
            "description": "内部错误",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "503": {
            "description": "服务暂不可用",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/monitoring/status": {
      "get": {
        "operationId": "get_monitoring_status",
        "tags": [
          "Monitoring"
        ],
        "summary": "查询worker与外部数据源状态",
        "description": "查询worker与外部数据源状态",
        "x-phase": "B0",
        "x-implementation-status": "planned",
        "x-permissions": [
          "admin"
        ],
        "security": [
          {
            "UserBearer": []
          }
        ],
        "parameters": [],
        "responses": {
          "200": {
            "description": "成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MonitoringStatus"
                },
                "example": {
                  "worker_status": "RUNNING",
                  "worker_heartbeat_at": "2026-09-06T00:00:00Z",
                  "due_job_count": 0,
                  "oldest_due_seconds": 0,
                  "sources": [
                    {
                      "source": "simulation",
                      "last_success_at": "2026-09-06T00:00:00Z",
                      "last_sequence": 1,
                      "status": "FRESH",
                      "last_error": null
                    }
                  ],
                  "agent_enabled": false,
                  "forecast_provider": "fixed"
                }
              }
            }
          },
          "401": {
            "description": "无有效身份",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "403": {
            "description": "无对应权限",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "资源不存在或不可见",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "422": {
            "description": "字段或业务约束无效",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "429": {
            "description": "超过频率限制",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "500": {
            "description": "内部错误",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "503": {
            "description": "服务暂不可用",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "UserBearer": {
        "type": "http",
        "scheme": "bearer",
        "description": "用户身份；角色与门店范围由服务端验证。"
      },
      "ServiceBearer": {
        "type": "http",
        "scheme": "bearer",
        "description": "服务身份；Agent无审批权限，simulation仅提交授权事件。"
      }
    },
    "schemas": {
      "ErrorDetail": {
        "type": "object",
        "properties": {
          "code": {
            "type": "string"
          },
          "message": {
            "type": "string"
          },
          "request_id": {
            "type": "string",
            "description": "不透明资源 ID",
            "minLength": 1,
            "maxLength": 128
          },
          "retryable": {
            "type": "boolean"
          },
          "details": {
            "type": "object",
            "additionalProperties": true
          }
        },
        "required": [
          "code",
          "message",
          "request_id",
          "retryable",
          "details"
        ],
        "additionalProperties": false
      },
      "Error": {
        "type": "object",
        "properties": {
          "error": {
            "$ref": "#/components/schemas/ErrorDetail"
          }
        },
        "required": [
          "error"
        ],
        "additionalProperties": false
      },
...
      "AlertFacts": {
        "type": "object",
        "properties": {
          "shortage_qty": {
            "type": "integer",
            "minimum": 0
          },
          "cash_floor_minor": {
            "type": "integer",
            "minimum": 0
          },
          "available_cash_minor": {
            "type": "integer",
            "minimum": 0
          },
          "action_id": {
            "type": "string",
            "description": "不透明资源 ID",
            "minLength": 1,
            "maxLength": 128
          },
          "data_as_of": {
            "type": "string",
            "description": "RFC3339 UTC 时间",
            "format": "date-time"
          },
          "reason": {
            "type": "string"
          }
        },
        "required": [],
        "additionalProperties": false
      },
      "Alert": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "不透明资源 ID",
            "minLength": 1,
            "maxLength": 128
          },
          "store_id": {
            "type": "string",
            "description": "不透明资源 ID",
            "minLength": 1,
            "maxLength": 128
          },
          "mission_id": {
            "anyOf": [
              {
                "type": "string",
                "description": "不透明资源 ID",
                "minLength": 1,
                "maxLength": 128
              },
              {
                "type": "null"
              }
            ]
          },
          "sku_id": {
            "anyOf": [
              {
                "type": "string",
                "description": "不透明资源 ID",
                "minLength": 1,
                "maxLength": 128
              },
              {
                "type": "null"
              }
            ]
          },
          "type": {
            "type": "string",
            "enum": [
              "STOCKOUT_RISK",
              "CASH_CONSTRAINT",
              "ACTION_EXCEPTION",
              "DATA_STALE"
            ]
          },
          "severity": {
            "type": "string",
            "enum": [
              "INFO",
              "WARNING",
              "CRITICAL"
            ]
          },
          "status": {
            "type": "string",
            "enum": [
              "OPEN",
              "ACKNOWLEDGED",
              "RESOLVED"
            ]
          },
          "summary": {
            "type": "string"
          },
          "facts": {
            "$ref": "#/components/schemas/AlertFacts"
          },
          "state_version": {
            "anyOf": [
              {
                "type": "integer",
                "minimum": 1
              },
              {
                "type": "null"
              }
            ]
          },
          "related_plan_id": {
            "anyOf": [
              {
                "type": "string",
                "description": "不透明资源 ID",
                "minLength": 1,
                "maxLength": 128
              },
              {
                "type": "null"
              }
            ]
          },
          "first_seen_at": {
            "type": "string",
            "description": "RFC3339 UTC 时间",
            "format": "date-time"
          },
          "last_seen_at": {
            "type": "string",
            "description": "RFC3339 UTC 时间",
            "format": "date-time"
          },
          "resolved_at": {
            "anyOf": [
              {
                "type": "string",
                "description": "RFC3339 UTC 时间",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "required": [
          "id",
          "store_id",
          "mission_id",
          "sku_id",
          "type",
          "severity",
          "status",
          "summary",
          "facts",
          "state_version",
          "related_plan_id",
          "first_seen_at",
          "last_seen_at",
          "resolved_at"
        ],
        "additionalProperties": false
      },
      "Reference": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string",
            "enum": [
              "mission",
              "plan",
              "action",
              "event",
              "forecast",
              "artifact",
              "scenario",
              "store",
              "alert"
            ]
          },
          "id": {
            "type": "string",
            "description": "不透明资源 ID",
            "minLength": 1,
            "maxLength": 128
          },
          "version": {
            "type": "string"
          }
        },
        "required": [
          "type",
          "id"
        ],
        "additionalProperties": false
      },
...
      "JobResult": {
        "type": "object",
        "properties": {
          "check_status": {
            "type": "string",
            "enum": [
              "HEALTHY",
              "ANOMALY",
              "INCONCLUSIVE",
              "SKIPPED"
            ]
          },
          "summary": {
            "type": "string"
          },
          "references": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Reference"
            }
          },
          "input_state_version": {
            "type": "integer",
            "minimum": 1
          },
          "output_state_version": {
            "type": "integer",
            "minimum": 1
          }
        },
        "required": [
          "summary",
          "references"
        ],
        "additionalProperties": false
      },
      "JobRun": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "不透明资源 ID",
            "minLength": 1,
            "maxLength": 128
          },
          "mission_id": {
            "anyOf": [
              {
                "type": "string",
                "description": "不透明资源 ID",
                "minLength": 1,
                "maxLength": 128
              },
              {
                "type": "null"
              }
            ]
          },
          "job_type": {
            "type": "string",
            "enum": [
              "worker_probe",
              "sync_events",
              "check_mission",
              "execute_purchase",
              "reconcile_action",
              "check_freshness",
              "agent_followup",
              "initialize_scenario",
              "advance_scenario"
            ]
          },
          "trigger_source": {
            "type": "string",
            "enum": [
              "INTERVAL",
              "AT",
              "EVENT",
              "MANUAL",
              "APPROVAL"
            ]
          },
          "status": {
            "type": "string",
            "enum": [
              "READY",
              "RUNNING",
              "SUCCEEDED",
              "FAILED",
              "CANCELLED",
              "RETRY_WAIT"
            ]
          },
          "attempt_count": {
            "type": "integer",
            "minimum": 0
          },
          "scheduled_for": {
            "type": "string",
            "description": "RFC3339 UTC 时间",
            "format": "date-time"
          },
          "available_at": {
            "type": "string",
            "description": "RFC3339 UTC 时间",
            "format": "date-time"
          },
          "started_at": {
            "anyOf": [
              {
                "type": "string",
                "description": "RFC3339 UTC 时间",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ]
          },
          "finished_at": {
            "anyOf": [
              {
                "type": "string",
                "description": "RFC3339 UTC 时间",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ]
          },
          "result": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/JobResult"
              },
              {
                "type": "null"
              }
            ]
          },
          "error_code": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ]
          },
          "last_error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ]
          },
          "request_id": {
            "anyOf": [
              {
                "type": "string",
                "minLength": 1,
                "maxLength": 128
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "required": [
          "id",
          "mission_id",
          "job_type",
          "trigger_source",
          "status",
          "attempt_count",
          "scheduled_for",
          "available_at",
          "started_at",
          "finished_at",
          "result",
          "error_code",
          "last_error",
          "request_id"
        ],
        "additionalProperties": false
      },
      "SourceStatus": {
        "type": "object",
        "properties": {
          "source": {
            "type": "string"
          },
          "last_success_at": {
            "anyOf": [
              {
                "type": "string",
                "description": "RFC3339 UTC 时间",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ]
          },
          "last_sequence": {
            "type": "integer",
            "minimum": 0
          },
          "status": {
            "type": "string",
            "enum": [
              "FRESH",
              "STALE",
              "UNKNOWN"
            ]
          },
          "last_error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "required": [
          "source",
          "last_success_at",
          "last_sequence",
          "status",
          "last_error"
        ],
        "additionalProperties": false
      },
      "MonitoringStatus": {
        "type": "object",
        "properties": {
          "worker_status": {
            "type": "string",
            "enum": [
              "RUNNING",
              "STALE",
              "STOPPED",
              "UNKNOWN"
            ]
          },
          "worker_heartbeat_at": {
            "anyOf": [
              {
                "type": "string",
                "description": "RFC3339 UTC 时间",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ]
          },
          "due_job_count": {
            "type": "integer",
            "minimum": 0
          },
          "oldest_due_seconds": {
            "type": "integer",
            "minimum": 0
          },
          "sources": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/SourceStatus"
            }
          },
          "agent_enabled": {
            "type": "boolean"
          },
          "forecast_provider": {
            "type": "string",
            "enum": [
              "fixed",
              "http"
            ]
          }
        },
        "required": [
          "worker_status",
          "worker_heartbeat_at",
          "due_job_count",
          "oldest_due_seconds",
          "sources",
          "agent_enabled",
          "forecast_provider"
        ],
        "additionalProperties": false
      },
...
      "AlertList": {
        "type": "object",
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Alert"
            }
          },
          "next_cursor": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "required": [
          "items",
          "next_cursor"
        ],
        "additionalProperties": false
      },
...
    }
  }
}